Separate css and html config in scaffold
Refactor entry files
