+++ src/templates/components/head/index.js
@@ -6,14 +6,14 @@ const head = ({
     meta,
     viewport = 'width=device-width, initial-scale=1.0, shrink-to-fit=no',
     favicon = `/static/img/favicon.ico`,
-    css,
+    css = '/static/css/index.css',
     children
 }) => <head>
     <meta charset={charset} />
     <meta name="viewport" content={viewport} />
     <link rel="shortcut icon" href={favicon} />
     <title>{title}</title>
-    {css && <link rel="stylesheet" href={`/${css}`} />}
+    {css && <link rel="stylesheet" href={`${css}`} />}
     {children}
     {meta && meta.map(item => {
         if (item.name) return <meta name={item.name} content={item.content} />;
