+++ tools/webpack/entry/css.js
@@ -1,0 +1,1 @@
+require('../../../src/css/index.scss');
