+++ tools/webpack/entry/html.js
@@ -1,7 +1,6 @@
 const h = require(`preact`).h;
 const render = require('preact-render-to-string').render;
-const Html = require(`./default-html`);
-require('../../../src/css/index.scss');
+const Html = require(`./template`);
 
 module.exports = function(locals) {
     const bodyTemplate = require(`../../../src/templates/pages/${locals.path}`).default;
@@ -11,16 +10,16 @@ module.exports = function(locals) {
     }
     try {
         const body = bodyTemplate();
-        const assets = Object.keys(locals.webpackStats.compilation.assets);
-        const css = assets.filter(value => value.match(/\.css$/));
+        // const assets = Object.keys(locals.webpackStats.compilation.assets);
+        // const css = assets.filter(value => value.match(/\.css$/));
         const title = require(`../../../src/templates/pages/${locals.path}`).title || '';
         const meta = require(`../../../src/templates/pages/${locals.path}`).meta || '';
         return new Promise((resolve, reject) => {
             if (body.then) {
                 body.then(Res => {
-                    resolve(`<!DOCTYPE html>${render(<Html css={css} htmlBody={Res} title={title} meta={meta} />)}`);
+                    resolve(`<!DOCTYPE html>${render(<Html htmlBody={Res} title={title} meta={meta} />)}`);
                 });
-            } else resolve(`<!DOCTYPE html>${render(<Html css={css} htmlBody={body} title={title} meta={meta} />)}`);
+            } else resolve(`<!DOCTYPE html>${render(<Html htmlBody={body} title={title} meta={meta} />)}`);
         });
     } catch (e){
         return Promise.reject(`HTML render error: ${e}`);
